refactor: Extract header logic from ChatArea into ChatHeader component
- Moved all inline header JSX from ChatArea.tsx to ChatHeader.tsx
- Added state management for topic editing (isEditingTopic, editedTopic)
- Added props for all required callbacks and data
- Removed ~260 lines of inline JSX from ChatArea.tsx
- ChatHeader now handles:
  * Channel/private chat/server notices display
  * Topic editing with inline form
  * Search functionality
  * Notification toggle
  * Channel settings, invite user, rename buttons
  * Member list toggle
  * Channel avatar support with fallback

This makes ChatArea.tsx cleaner and ChatHeader.tsx more reusable.
All tests passing ✅
