+++ src/components/layout/ChatHeader.tsx
@@ -1,8 +1,10 @@
 import { UsersIcon } from "@heroicons/react/24/solid";
 import type React from "react";
+import { useState } from "react";
 import {
   FaAt,
   FaBell,
+  FaBellSlash,
   FaChevronLeft,
   FaChevronRight,
   FaEdit,
@@ -13,31 +15,57 @@ import {
   FaUserPlus,
 } from "react-icons/fa";
 import ircClient from "../../lib/ircClient";
+import {
+  getChannelAvatarUrl,
+  getChannelDisplayName,
+  hasOpPermission,
+} from "../../lib/ircUtils";
 import useStore from "../../store";
-import type { Channel, PrivateChat } from "../../types";
+import type { Channel, PrivateChat, User } from "../../types";
 
 interface ChatHeaderProps {
   selectedChannel: Channel | null;
   selectedPrivateChat: PrivateChat | null;
   selectedServerId: string | null;
+  selectedChannelId: string | null;
+  currentUser: User | null;
   isChanListVisible: boolean;
   isMemberListVisible: boolean;
   isNarrowView: boolean;
+  globalSettings: {
+    notificationVolume: number;
+  };
+  searchQuery: string;
   onToggleChanList: () => void;
   onToggleMemberList: () => void;
+  onSearchQueryChange: (query: string) => void;
+  onToggleNotificationVolume: () => void;
+  onOpenChannelSettings: () => void;
+  onOpenInviteUser: () => void;
 }
 
 export const ChatHeader: React.FC<ChatHeaderProps> = ({
   selectedChannel,
   selectedPrivateChat,
   selectedServerId,
+  selectedChannelId,
+  currentUser,
   isChanListVisible,
   isMemberListVisible,
   isNarrowView,
+  globalSettings,
+  searchQuery,
   onToggleChanList,
   onToggleMemberList,
+  onSearchQueryChange,
+  onToggleNotificationVolume,
+  onOpenChannelSettings,
+  onOpenInviteUser,
 }) => {
-  const { toggleChannelListModal, toggleChannelRenameModal } = useStore();
+  const { toggleChannelListModal, toggleChannelRenameModal, toggleMemberList } =
+    useStore();
+  const [isEditingTopic, setIsEditingTopic] = useState(false);
+  const [editedTopic, setEditedTopic] = useState("");
 
   // Check if current user is operator
   const isOperator = (() => {
@@ -54,24 +82,134 @@ export const ChatHeader: React.FC<ChatHeaderProps> = ({
   })();
 
   return (
-    <div className="h-12 min-h-[48px] px-4 border-b border-discord-dark-400 flex items-center justify-between shadow-sm">
-      <div className="flex items-center">
+    <div className="min-h-[56px] px-4 border-b border-discord-dark-400 flex flex-wrap items-start md:items-center justify-between shadow-sm py-2 md:py-0 md:h-12 gap-y-2">
+      <div className="flex items-center flex-1 min-w-0 w-full md:w-auto">
         {!isChanListVisible && (
           <button
             onClick={onToggleChanList}
-            className="text-discord-channels-default hover:text-white mr-4"
+            className="text-discord-channels-default hover:text-white mr-4 flex-shrink-0"
             aria-label="Expand channel list"
           >
             {isNarrowView ? <FaChevronLeft /> : <FaChevronRight />}
           </button>
         )}
         {selectedChannel && (
-          <>
-            <FaHashtag className="text-discord-text-muted mr-2" />
-            <h2 className="font-bold text-white mr-4">
-              {selectedChannel.name.replace(/^#/, "")}
-            </h2>
-          </>
+          <div className="flex flex-col min-w-0 flex-1 md:flex-row md:items-center">
+            <div className="flex items-center min-w-0 flex-shrink-0">
+              {getChannelAvatarUrl(selectedChannel.metadata, 50) ? (
+                <img
+                  src={getChannelAvatarUrl(selectedChannel.metadata, 50)}
+                  alt={selectedChannel.name}
+                  className="w-12 h-12 rounded-full object-cover mr-2 flex-shrink-0"
+                  onError={(e) => {
+                    // Fallback to # icon on error
+                    e.currentTarget.style.display = "none";
+                    const parent = e.currentTarget.parentElement;
+                    const fallbackIcon = parent?.querySelector(
+                      ".fallback-hash-icon",
+                    );
+                    if (fallbackIcon) {
+                      (fallbackIcon as HTMLElement).style.display =
+                        "inline-block";
+                    }
+                  }}
+                />
+              ) : null}
+              <FaHashtag
+                className="text-discord-text-muted mr-2 fallback-hash-icon flex-shrink-0 text-3xl"
+                style={{
+                  display: getChannelAvatarUrl(selectedChannel.metadata, 50)
+                    ? "none"
+                    : "inline-block",
+                }}
+              />
+              <h2 className="font-bold text-white mr-4 truncate">
+                {getChannelDisplayName(
+                  selectedChannel.name,
+                  selectedChannel.metadata,
+                )}
+              </h2>
+            </div>
+            <div className="md:mx-2 md:text-discord-text-muted hidden md:block">
+              |
+            </div>
+            {isEditingTopic ? (
+              <form
+                onSubmit={(e) => {
+                  e.preventDefault();
+                  if (selectedServerId && selectedChannel) {
+                    ircClient.setTopic(
+                      selectedServerId,
+                      selectedChannel.name,
+                      editedTopic,
+                    );
+                    setIsEditingTopic(false);
+                  }
+                }}
+                className="flex items-center gap-2 flex-1 max-w-md mt-1 md:mt-0"
+              >
+                <input
+                  type="text"
+                  value={editedTopic}
+                  onChange={(e) => setEditedTopic(e.target.value)}
+                  onKeyDown={(e) => {
+                    if (e.key === "Escape") {
+                      setIsEditingTopic(false);
+                    }
+                  }}
+                  autoFocus
+                  className="flex-1 px-2 py-1 bg-discord-dark-300 text-white text-sm rounded"
+                  placeholder="Enter topic..."
+                />
+                <button
+                  type="submit"
+                  className="px-3 py-1 bg-discord-primary hover:bg-opacity-80 text-white text-sm rounded"
+                >
+                  Save
+                </button>
+                <button
+                  type="button"
+                  onClick={() => setIsEditingTopic(false)}
+                  className="px-3 py-1 bg-discord-dark-400 hover:bg-discord-dark-500 text-white text-sm rounded"
+                >
+                  Cancel
+                </button>
+              </form>
+            ) : (
+              <button
+                onClick={() => {
+                  if (selectedChannel) {
+                    const currentUserInChannel = selectedChannel.users.find(
+                      (u) => u.username === currentUser?.username,
+                    );
+                    if (hasOpPermission(currentUserInChannel?.status)) {
+                      setEditedTopic(selectedChannel.topic || "");
+                      setIsEditingTopic(true);
+                    }
+                  }
+                }}
+                className="text-discord-text-muted text-xs md:text-sm hover:text-white truncate min-w-0 md:max-w-md mt-0.5 mb-1 md:mt-0 md:mb-0"
+                title={
+                  selectedChannel.topic
+                    ? `Topic: ${selectedChannel.topic}${
+                        selectedChannel.users.find(
+                          (u) => u.username === currentUser?.username,
+                        ) &&
+                        hasOpPermission(
+                          selectedChannel.users.find(
+                            (u) => u.username === currentUser?.username,
+                          )?.status,
+                        )
+                          ? " (Click to edit)"
+                          : ""
+                      }`
+                    : "No topic set"
+                }
+              >
+                {selectedChannel.topic || "No topic"}
+              </button>
+            )}
+          </div>
         )}
         {selectedPrivateChat && (
           <>
@@ -81,26 +219,48 @@ export const ChatHeader: React.FC<ChatHeaderProps> = ({
             </h2>
           </>
         )}
-        {selectedChannel?.topic && (
+        {selectedChannelId === "server-notices" && (
           <>
-            <div className="mx-2 text-discord-text-muted">|</div>
-            <div className="text-discord-text-muted text-sm truncate max-w-xs">
-              {selectedChannel.topic}
-            </div>
+            <FaList className="text-discord-text-muted mr-2" />
+            <h2 className="font-bold text-white mr-4">Server Notices</h2>
           </>
         )}
       </div>
-      {!!selectedServerId && (
-        <div className="flex items-center gap-4 text-discord-text-muted">
-          <button className="hover:text-discord-text-normal">
-            <FaBell />
+      {!!selectedServerId && selectedChannelId !== "server-notices" && (
+        <div className="flex items-center gap-2 md:gap-4 text-discord-text-muted flex-shrink-0">
+          <button
+            className="hover:text-discord-text-normal"
+            onClick={onToggleNotificationVolume}
+            title={
+              globalSettings.notificationVolume > 0
+                ? "Mute notifications"
+                : "Enable notifications"
+            }
+          >
+            {globalSettings.notificationVolume > 0 ? (
+              <FaBell />
+            ) : (
+              <FaBellSlash />
+            )}
           </button>
-          <button className="hover:text-discord-text-normal">
-            <FaPenAlt />
-          </button>
-          <button className="hover:text-discord-text-normal">
-            <FaUserPlus />
-          </button>
+          {selectedChannel && (
+            <button
+              className="hover:text-discord-text-normal"
+              onClick={onOpenChannelSettings}
+              title="Channel Settings"
+            >
+              <FaPenAlt />
+            </button>
+          )}
+          {selectedChannel && (
+            <button
+              className="hover:text-discord-text-normal"
+              onClick={onOpenInviteUser}
+              title="Invite User"
+            >
+              <FaUserPlus />
+            </button>
+          )}
           <button
             className="hover:text-discord-text-normal"
             onClick={() => toggleChannelListModal(true)}
@@ -121,7 +281,7 @@ export const ChatHeader: React.FC<ChatHeaderProps> = ({
           {selectedChannel && (
             <button
               className="hover:text-discord-text-normal"
-              onClick={() => onToggleMemberList()}
+              onClick={() => toggleMemberList(!isMemberListVisible)}
               aria-label={
                 isMemberListVisible
                   ? "Collapse member list"
@@ -140,10 +300,31 @@ export const ChatHeader: React.FC<ChatHeaderProps> = ({
             <input
               type="text"
               placeholder="Search"
-              className="bg-discord-dark-400 text-discord-text-muted text-sm rounded px-2 py-1 w-32 focus:outline-none focus:ring-1 focus:ring-discord-text-link"
+              value={searchQuery}
+              onChange={(e) => onSearchQueryChange(e.target.value)}
+              className="bg-discord-dark-400 text-discord-text-muted text-sm rounded px-2 py-1 w-20 md:w-32 focus:outline-none focus:ring-1 focus:ring-discord-text-link"
             />
             <FaSearch className="absolute right-2 top-1.5 text-xs" />
           </div>
+        </div>
+      )}
+      {selectedChannelId === "server-notices" && (
+        <div className="flex items-center gap-4 text-discord-text-muted">
+          {/* TODO: Re-enable pop out button for server notices
+          <button
+            className="hover:text-discord-text-normal"
+            onClick={() =>
+              setIsServerNoticesPoppedOut(!isServerNoticesPoppedOut)
+            }
+            title={
+              isServerNoticesPoppedOut
+                ? "Pop in server notices"
+                : "Pop out server notices"
+            }
+          >
+            <FaExternalLinkAlt />
+          </button>
+          */}
         </div>
       )}
     </div>
